fix: address code review comments
